backend/requirements.txt
fastapi
pydantic
uvicorn
torch==2.6.0  # Specific version that works with drawatoon on M3 Max
pillow==9.0.0
tqdm==4.62.3
numpy==1.24.3
pathlib==1.0.1
diffusers
transformers
einops
sentencepiece
opencv-python
beautifulsoup4